refactor record hydration to RecordSet class
