+++ src/Glpi/Debug/Profiler.php
@@ -55,6 +55,7 @@ final class Profiler
     public const CATEGORY_TWIG = 'twig';
     public const CATEGORY_SEARCH = 'search';
     public const CATEGORY_CUSTOMOBJECTS = 'customobjects';
+    public const CATEGORY_HLAPI = 'hlapi';
 
     private static $instance;
